split models into different files
