+++ mcodingbot/database/models/__init__.py
@@ -1,0 +1,6 @@
+from typing import Sequence
+from mcodingbot.database.models.user import User
+from mcodingbot.database.models.word import Word
+from mcodingbot.database.models.user_word import UserWord
+
+__all__: Sequence[str] = ("User", "Word", "UserWord")
